refact, fix: move keystore generation with fixes and yubikey setup to api
Fix generation of keystore files. Update generation of delegated roles'
keystore files and update TUF call
