+++ taf/api/yubikey.py
@@ -1,0 +1,57 @@
+import click
+
+from pathlib import Path
+from taf.constants import DEFAULT_RSA_SIGNATURE_SCHEME
+import taf.yubikey as yk
+
+
+def export_yk_public_pem(path=None):
+    try:
+        pub_key_pem = yk.export_piv_pub_key().decode("utf-8")
+    except Exception:
+        print("Could not export the public key. Check if a YubiKey is inserted")
+        return
+    if path is None:
+        print(pub_key_pem)
+    else:
+        if not path.endswith(".pub"):
+            path = f"{path}.pub"
+        path = Path(path)
+        parent = path.parent
+        parent.mkdir(parents=True, exist_ok=True)
+        path.write_text(pub_key_pem)
+
+
+def setup_signing_yubikey(certs_dir=None, scheme=DEFAULT_RSA_SIGNATURE_SCHEME):
+    if not click.confirm(
+        "WARNING - this will delete everything from the inserted key. Proceed?"
+    ):
+        return
+    _, serial_num = yk.yubikey_prompt(
+        "new Yubikey",
+        creating_new_key=True,
+        pin_confirm=True,
+        pin_repeat=True,
+        prompt_message="Please insert the new Yubikey and press ENTER",
+    )
+    key = yk.setup_new_yubikey(serial_num)
+    yk.export_yk_certificate(certs_dir, key)
+
+
+def setup_test_yubikey(key_path=None):
+    """
+    Resets the inserted yubikey, sets default pin and copies the specified key
+    onto it.
+    """
+    if not click.confirm("WARNING - this will reset the inserted key. Proceed?"):
+        return
+    key_path = Path(key_path)
+    key_pem = key_path.read_bytes()
+
+    print(f"Importing RSA private key from {key_path} to Yubikey...")
+    pin = yk.DEFAULT_PIN
+
+    pub_key = yk.setup(pin, "Test Yubikey", private_key_pem=key_pem)
+    print("\nPrivate key successfully imported.\n")
+    print("\nPublic key (PEM): \n{}".format(pub_key.decode("utf-8")))
+    print("Pin: {}\n".format(pin))
